django_app/requirements.txt
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dj-database-url==0.5.0
psycopg2==2.8.4
